+++ database/migrations/2024_06_02_000002_create_property_status_histories_table.php
@@ -8,14 +8,16 @@ return new class extends Migration
 {
     public function up()
     {
-        Schema::create('property_status_histories', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('property_id')->constrained()->onDelete('cascade');
-            $table->foreignId('status_id')->constrained('statuses');
-            $table->timestamp('start_date')->useCurrent();
-            $table->timestamp('end_date')->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable("property_status_histories")) {
+            Schema::create('property_status_histories', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('property_id')->constrained()->onDelete('cascade');
+                $table->foreignId('status_id')->constrained('statuses');
+                $table->timestamp('start_date')->useCurrent();
+                $table->timestamp('end_date')->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down()
